fixed: MOD with negative numbers. the two operands have to be signed. y_fluchtpunkt had a wrong value from refline_settings. We now use rowPerspectivePx.

+++ traktor/src/pipeline_2_detect.cpp
@@ -248,7 +248,7 @@ bool is_within_threshold(const float avg_threshold, const int rowSpacingPx, cons
 */
 
 
-HARROW_DIRECTION get_harrow_direction(const bool is_within_threshold, const float avg_threshold)
+HARROW_DIRECTION get_harrow_direction(const bool is_within_threshold, const int avg_threshold)
 {
     HARROW_DIRECTION direction;
